# backend/api/swagger.yaml
openapi: 3.0.3
info:
  title: PUN street Universal Access - OpenAPI 3.0
  description: pua
  version: v1
servers:
- url: http://localhost:5000/api/v1
paths:
...
  /user/get-info/{userID}:
    get:
      tags:
      - user
      summary: get user info by ID
      parameters:
      - name: userID
        in: path
        description: The ID of the user to return.
        required: true
        content:
          application/json:
            schema:
              type: string
              example: "1"
              x-content-type: application/json
      responses:
        "200":
          description: return user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserData'
        "400":
          description: bad request
        "500":
          description: Unexpected error
  /login:
    post:
      tags:
      - user
      summary: login
      requestBody:
        description: login info
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInfo'
      responses:
        "200":
          description: login success
        "400":
          description: bad request
        "500":
          description: Unexpected error
  /validate:
    get:
      tags: 
      - user
      summary: validate jwt token
      responses:
        "200":
          description: validate success
        "401":
          description: Unauthorized
        "500":
          description: Unexpected error
...
  /admin/get-all-users:
    get:
      tags:
      - admin
      - user
      summary: get all users
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserArray'
        "400":
          description: bad request
        "500":
          description: Unexpected error
...
components:
  schemas:
...
    UserData:
      required:
      - address
      - authority
      - birthday
      - cart_id
      - password
      - phone
      - status
      - user_email
      - user_id
      - user_name
      type: object
      properties:
        user_id:
          type: integer
          format: int64
          example: 1
        user_name:
          type: string
          example: pasta
        user_email:
          type: string
          example: t110595959@rice.org.tw
        authority:
          type: integer
          format: int64
          example: 100
        password:
          type: string
          example: "123456"
        address:
          type: string
          example: pun street
        phone:
          type: string
          example: "0912345678"
        status:
          type: integer
          format: int64
          example: 1
        cart_id:
          type: integer
          format: int64
          example: 1
        birthday:
          type: string
          example: 2021-05-01
      example:
        birthday: 2021-05-01
        cart_id: 1
        user_email: t110595959@rice.org.tw
        password: "123456"
        address: pun street
        user_id: 1
        phone: "0912345678"
        user_name: pasta
        authority: 100
        status: 1
    UserArray:
      type: array
      items:
        $ref: '#/components/schemas/UserData'
      x-schema-name: UserArray
...
    LoginInfo:
      required:
      - password
      - user_email
      properties:
        user_email:
          type: string
          example: a@gmail.com
        password:
          type: string
          example: pwd1
...